eval/pyproject.toml
[project]
name = "eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "inspect-ai>=0.3.217",
    "jsonschema>=4.26.0",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.3",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
